# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema Pesquisas — API Multi-Tenant",
        "description": "Backend REST API multiempresa com RBAC. Use POST /api/v1/auth/login para obter o JWT e clique em Authorize inserindo: Bearer {token}. Super Admins podem usar o header X-Empresa-Id para escopar requisições a uma empresa.",
        "contact": {
            "email": "admin@sistema.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/configuracoes": {
            "get": {
                "tags": [
                    "Configurações"
                ],
                "summary": "Listar configurações",
                "operationId": "b99cde0857a23a856b6d22de08ae5c5f",
                "parameters": [
                    {
                        "name": "grupo",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Configurações",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sucesso": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "dados": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Configurações"
                ],
                "summary": "Salvar configuração (upsert)",
                "operationId": "0126c60ac9bb0e6ef489a3b604bb4d1c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "chave",
                                    "valor"
                                ],
                                "properties": {
                                    "chave": {
                                        "type": "string",
                                        "example": "smtp.host"
                                    },
                                    "valor": {
                                        "type": "string",
                                        "example": "smtp.gmail.com"
                                    },
                                    "tipo": {
                                        "type": "string",
                                        "enum": [
                                            "string",
                                            "boolean",
                                            "integer",
                                            "json"
                                        ]
                                    },
                                    "grupo": {
                                        "type": "string",
                                        "example": "email"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Salvo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sucesso": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/configuracoes/{chave}": {
            "get": {
                "tags": [
                    "Configurações"
                ],
                "summary": "Buscar configuração por chave",
                "operationId": "f02bdfd2904c66636bf346c66c872413",
                "parameters": [
                    {
                        "name": "chave",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "smtp.host"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Configuração encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sucesso": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "dados": {
                                            "$ref": "#/components/schemas/Configuracao"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Chave não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RespostaErro"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Configurações"
                ],
                "summary": "Remover configuração",
                "operationId": "6c0b2330c764230e53e634f261a3079c",
                "parameters": [
                    {
                        "name": "chave",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Removida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sucesso": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/configuracoes/lote": {
            "post": {
                "tags": [
                    "Configurações"
                ],
                "summary": "Salvar múltiplas configurações",
                "operationId": "fa810448a513822a4eff076255913ea8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "configuracoes"
                                ],
                                "properties": {
                                    "configuracoes": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "chave": {
                                                    "type": "string"
                                                },
                                                "valor": {
                                                    "type": "string"
                                                },
                                                "grupo": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Salvas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sucesso": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "mensagem": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "RespostaErro": {
                "description": "Resposta de erro padrão da API",
                "properties": {
                    "sucesso": {
                        "type": "boolean",
                        "example": false
                    },
                    "mensagem": {
                        "type": "string",
                        "example": "Mensagem descritiva do erro."
                    }
                },
                "type": "object"
            },
            "Configuracao": {
                "description": "Configuração chave-valor por empresa/filial",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "empresa_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "filial_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "chave": {
                        "type": "string",
                        "example": "smtp.host"
                    },
                    "valor": {
                        "type": "string",
                        "example": "smtp.gmail.com",
                        "nullable": true
                    },
                    "tipo": {
                        "type": "string",
                        "example": "string",
                        "enum": [
                            "string",
                            "boolean",
                            "integer",
                            "json"
                        ]
                    },
                    "grupo": {
                        "type": "string",
                        "example": "email",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token JWT obtido via POST /api/v1/auth/login. Insira: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}